Vision restructure for simulator
