+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/vision/CameraFunctions.java
@@ -1,0 +1,46 @@
+package org.firstinspires.ftc.teamcode.vision;
+
+import com.acmerobotics.dashboard.FtcDashboard;
+import com.qualcomm.robotcore.hardware.HardwareMap;
+
+
+import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
+import org.firstinspires.ftc.teamcode.TeleOpConfig;
+import org.opencv.core.Core;
+import org.opencv.core.Mat;
+import org.opencv.core.Point;
+import org.opencv.core.Rect;
+import org.opencv.core.Scalar;
+import org.opencv.engine.OpenCVEngineInterface;
+import org.opencv.imgproc.Imgproc;
+import org.openftc.easyopencv.OpenCvCamera;
+import org.openftc.easyopencv.OpenCvCameraFactory;
+import org.openftc.easyopencv.OpenCvCameraRotation;
+import org.openftc.easyopencv.OpenCvPipeline;
+
+public class CameraFunctions {
+
+    public HardwareMap hardwareMap;
+    public OpenCvCamera extCam;
+
+
+    public void initVision (HardwareMap hw, OpenCvPipeline chosenPipeline) {
+        this.hardwareMap = hw;
+        int cameraMonitorViewId = hardwareMap.appContext.getResources().getIdentifier("cameraMonitorViewId", "id", hardwareMap.appContext.getPackageName());
+        WebcamName webcamName = hardwareMap.get(WebcamName.class,"webcam");
+        extCam = OpenCvCameraFactory.getInstance().createWebcam(webcamName, cameraMonitorViewId);
+
+
+        extCam.setPipeline(chosenPipeline);
+        extCam.openCameraDeviceAsync(new OpenCvCamera.AsyncCameraOpenListener()
+        {
+            @Override
+            public void onOpened()
+            {
+                extCam.startStreaming(320,240, OpenCvCameraRotation.SIDEWAYS_LEFT);
+                FtcDashboard.getInstance().startCameraStream(extCam,30);
+            }
+        });
+    }
+
+}
